refactor: FileReader 패키지 이동
