+++ tomcat/src/main/java/nextstep/jwp/presentation/RegisterController.java
@@ -2,7 +2,7 @@ package nextstep.jwp.presentation;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.coyote.http11.request.FileReader;
+import org.apache.coyote.http11.response.FileReader;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
