+++ tomcat/src/test/java/org/apache/coyote/http11/response/HttpResponseTest.java
@@ -1,6 +1,5 @@
 package org.apache.coyote.http11.response;
 
-import org.apache.coyote.http11.request.FileReader;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
